+++ vnext/Microsoft.ReactNative.Cxx/JSI/NodeApiJsiRuntime.h
